\name{ppc-internal}
\title{Internal ppc functions}
\alias{ppc.make.centroids}
\alias{which.is.min}
\alias{medoid}
\alias{permute.rows}
\alias{balanced.folds}
\alias{hclust.1d}
\alias{ppc.read.peaks.file}
\description{Internal ppc functions}
\usage{
ppc.make.centroids(clust.tree, x, user.parms)
which.is.min(x)
medoid(x)
permute.rows(x)
balanced.folds(y, nfolds)
hclust.1d(x, debug=FALSE)
ppc.read.peaks.file(filename)
}
\author{Balasubramanian Narasimhan
 and Rob Tibshirani}
\details{
  These are not to be called by the user.
}
\keyword{internal}

