\name{ppc.plotcv}
\alias{ppc.plotcv}
\title{ Function to plot CV curves from PPC analysis }
\description{
Function to plot CV curves from PPC analysis 
}
\usage{
ppc.plotcv(fit)
}
\arguments{
  \item{fit}{ Result of call to ppc.cv }
}
\details{
}
\value{
}
\author{ Balasubramanian Narasimhan and Rob Tibshirani }


\examples{
## for a complete worked example of this function in a PPC analysis  see
##  http://www-stat.stanford.edu/~tibs/PPC/Rdist/Rscript.rawdata
}

\keyword{ }% at least one, from doc/KEYWORDS


