\name{ppc.plotcvprob}
\alias{ppc.plotcvprob}
\title{ Function to plot CV probabilities from PPC analysis }
\description{
 Function to plot CV probabilities from PPC analysis
}
\usage{
ppc.plotcvprob(fit, data, threshold)
}
\arguments{
  \item{fit}{ Result of call to ppc.cv }
  \item{data}{ List containing mass spec data}
  \item{threshold}{ Value of threshold to use for predictions}
}
\details{
}
\value{
}
\author{ Balasubramanian Narasimhan and Rob Tibshirani }


\examples{
## for a complete worked example of this function in a PPC analysis  see
##  http://www-stat.stanford.edu/~tibs/PPC/Rdist/Rscript.rawdata
}

\keyword{ }% at least one, from doc/KEYWORDS


