\name{ppc.plotfdr}
\alias{ppc.plotfdr}
\title{ Function to plot FDR results from PPC analysis}
\description{
Function to plot FDR results from PPC analysis
}
\usage{
ppc.plotfdr(fdrfit)
}
\arguments{
  \item{fdrfit}{Result of call to ppc.fdr}
}
\details{
}
\value{
}
\author{ Balasubramanian Narasimhan and Rob Tibshirani }


\examples{
## for a complete worked example of this function in a PPC analysis  see
##  http://www-stat.stanford.edu/~tibs/PPC/Rdist/Rscript.rawdata
}

\keyword{ }% at least one, from doc/KEYWORDS


