\name{ppc.remove.suffix}
\alias{ppc.remove.suffix}
\title{ Function to remove the suffix from a file name }
\description{
This function just removes a .xxx suffix from the end of a filename
 or a vector of filenames
}
\usage{
ppc.remove.suffix(x)
}
\arguments{
  \item{x}{ Filename or vector of filenames }
}
\value{
Filename or list of filenames with suffix removed
}
\author{ Balasubramanian Narasimhan and Rob Tibshirani }
\examples{
## for a complete worked example of this function in a PPC analysis  see
##  http://www-stat.stanford.edu/~tibs/PPC/Rdist/Rscript.rawdata
}
\keyword{}

